requirements.txt
asgiref==3.2.7
Django==3.0.4
django-appconf==1.0.3
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-multiselectfield==0.1.12
django-organizations==1.1.2
django-select2==7.2.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
